moved writer unit tests to integration and fixed geoadd issues
